Fix every warnings

+++ src/services.c
@@ -5,16 +5,18 @@
 
 #include "services.h"
 
-void route(Request *request) {
+void route(Request *request)
+{
 
-    if(strcmp(request->uri, "/home") == 0 ) {
+    if(strcmp(request->uri, "/home") == 0 )
+    {
 
         Response *resp = http_create_response(request->clientfd);
         resp->response_code = 200;
 
         const char* hello = "Hello World ! ";
         resp->content.content = malloc(strlen(hello));
-        resp->content.content_length = strlen(hello);
+        resp->content.content_length = (int)strlen(hello);
         strcpy(resp->content.content, hello);
 
         http_send_response(request, resp);
@@ -29,7 +31,7 @@ void route(Request *request) {
 
     const char* notfound = "404 Page Not Found";
     resp->content.content = malloc(strlen(notfound));
-    resp->content.content_length = strlen(notfound);
+    resp->content.content_length = (int)strlen(notfound);
     strcpy(resp->content.content, notfound);
 
     http_send_response(request, resp);
